requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
bcrypt==3.2.2
cffi==1.15.0
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.1.1
psycopg2-binary==2.9.3
pycparser==2.21
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.2.0
